+++ examples/esp32/apps/memfault_demo_app/main/button.h
@@ -1,0 +1,10 @@
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+//!
+//! Button setup and handling
+
+#pragma once
+
+void button_setup(void);